backend event queue (#14)
An event queue is created to notify changes (e.g., node disconnection)
to the channel manager so that the channel manager can update that
information for each affected channel.

In addition to that, a few bugs are fixed. In particular, some
variables are accessed across two different asyncio loops, which caused
race condition.

Codes are also refactored and cleaned up to improve readability.