openapi spec update: design code API
REST APIs for design code retrieval and update are added.
New codes are generated based on the revised spec.
However, the generated codes have two bugs; one in api_design_code.go
and the other in api_designs.go. Every time they are newly generated,
these need to be fixed manually. We need a better way to handle this.