fix for mqtt client termination issue (#72)
There is a bug in mqtt backend implemtnation where a subscriber
fails to receive a message from a publisher when the publisher terminates.
This is due to the fact that mqtt last will message is used to signal
the termination of a node, which is an out-of-band mechanism.

When a publisher terminates, the subscriber now monitors if there is
any incoming message or not with a timeout deadline. Only after a
timeout, the allocated resources for the publisher is removed.

Also, publish() returns MQTTMessageInfo class, which has a function called
wait_for_publish(). This function call is blocked until publish is completed.
herefore, it ensures the completion of publication.