remove trailing slash in the rest api endpoints (#83)
A trailing slash in an rest api endpoint is generating a 301
permanently moved error code in some case. For example, ngnix somehow
removed the trailing slash in {user}/designs/, and since apiserver in
gorilla mux has StrictSlash set to true, it returns the 301 error and
the client turns POST request into GET request.

To mitigate the issue, the trailing slash is removed from all the
endpoints as the trailing slash doesn't have any meaningful purpose.