refactor fledgectl command structures (#97)
The fledgectl's command structure is updated with the following
format: fledgectl <verb> <resource> <name> [flags]. This improves the
usability of the tool. Along the way, openapi endpoints are also
updated to be more consistent and to reduce confusions. As a result,
major code refactoring took place across all openapi-related codebase.