refactor job api in openapi specification (#108)
Job API is revised to separate scheduling a job from specifying it.
More specifically, CreateJob, DeleteJob, GetJob and UpdateJob are for
taking actions on a job's specification. GetJobStatus and
GetJobsStatus are for obtaining job's status. UpdateJobStatus only
accepts "deploying" and "stopping" as its request data and supports
the scheduling of a job.

api_job.go and api_job_service.go are backed as .old files. They are
to be removed once refactoring job api is completed.