partial implementation of starting a job (#113)
The feature of starting a job is partially implemented. Once the
"start job" command is triggered by using fledgectl tool, the request
eventually reaches the controller via the apiserver. The controller
takes the status update request (i.e., start job command) and push the
request to a job manager subsystem. The job manager subsystem is not
yet completely implemented; it simply returns an error. The
implementation of the job manager will be conducted next.