save job payloads in the db (#116)
The payloads of a job are saved in the db so that fledgelet can fetch
its corresponding payload. To enable that, when payload is generated,
agent id is created and associated with the payload. Then, it is
stored in the db.

When a compute node is created, an agent id is configured in each node
so that fledgelet knows what payload it needs to fetch.