notify job to fledgelet (#118)
When scheduling a job is submitted (i.e. fledgectl start job), that
event should be distributed to all fledgelet agents that will be
participating in the job. The notification mechanism is revised and
simplified.

The old code is only commented out at the moment; it will be cleaned
up in the future revisions.