bugfix: too many connection attempts by fledglet (#124)
Somehow if a connection to notifier gets disconnected right after
connection is made, this can cause too many attempts to connect to the
notifier from fledgelet. This not only creates overheads to notifier
but also produces too many logs in the system, (thus creating
overheads to a monitoring system).

To mitigate the issue, static pause time (10 sec for now) is added in
the loop.