python lib compatibility (#128)
So far python library is not compatible with fledge control plane; in
particular, the library's configuration is different from the one
produced by the fledge control plane. This is now addressed.

To test that, fiab environment is updated. Now the mnist example in
the examples folder is executed in the fiab environment.

In addition, the on_message routine that uses asyncio future routine
has performance issue and bug. The bug is now fixed by using deque.