support for multiple jobs with helm template (#134)
helm chart is created on a per-job basis. This means that heml chart
files (e.g., Chart.yaml and values.yaml) should be copied into a
folder (e.g., job folder) and templates folder should be created
within in that folder. This way multiple jobs can be scheduled
in parallel.