resource spec update for job (#13)
A machine learning (ML) workload tends to need more compute resources.
Especially, memory limit often kicks off OOM, and leads to the
termination of an ML process. If requested memory is too much,
launching the workload can fail too.

This is an attempt to strike a balance between request and limit.
In future, request and limit will be dynamically configurable.