mlflow support
In order to save metrics and model from an FL job in a model registry,
mlflow client is integrated wtih the fledge library.

While doing so, the folder structures are updated such that all the
files in the registry folder are moved to "discovery" folder.

In the future, other model registries (e.g., neptune) may be
considered.

In addition, openapi specification is updated to include funcTags in
channel schema and basedModel object in job schema.