usecase: support user-driven compute and dataset
Participants joining an FL training job may not want to register their
system (e.g. k8s) into the fledge system. Instead, they may want to
join the FL job by launching a compute node (where their dataset is
based) under their control.

In order to support such a usecase, a dataset concept is revised and
two new properties (fromUser and fromSystem) are introduced in the
dataset.

fromUser property contains key-value pairs where key is a realm and
value is the number of datasets within the realm.

fromSystem property contains a list of dataset IDs that the fledge
system maintains.

When fromUser property is specified, the owner of an FL job should be
able to distribute agent IDs to their respective participants. Then,
the participants should use a container image containing fledgelet
agent, and set an environment variable (i.e., FLEDGE_AGENT_ID) when a
container is spun up. The way to bring up a container is the
responsibility of the participants.

In case of fromSystem, the fledge system figures out how to provision
compute resources and datasets.