fix: silent failure in initializing mqtt client connection
In the mqtt backend implementation, a failure to client connection was
not handled at all. Also, a timeout value is set too small in the
coroutine for the mqtt client connection. Hence, from time to time,
connection was silently failed.

A timeout is set to 10 sec and when timeout happens, the code now
raises a connection exception.