make image configurable
image is now specified as a controller's config parameter. The
specified image is used when a job is scheduled.

A future plan is to provide a set of images so that the controller can
dynamically choose images depending on resource availability (e.g.,
cpu, gpu, etc) of nodes in a cluster.

minikube dns issue resolution was reverted to a previous version
because the newer resolution causes domain name resolution failure in
pods.