composer loop exit logic update (#59)
loop exit condition is only tested in the last tasklet of a loop. It's
quite inflexible and limits the composer's functionality.

Now a condition is checked at every tasklet in a loop and we can get
out of the loop at any tasklet in the loop as soon as loop exit
conditions are met.

Therefore, when a role is composed, user needs to decide where the
exit conditions need to be updated.