feature: add random selector
Random selector is used in FedAvg, and it plays an important role to reduce the communication overhead while there are many local clients/trainers. This will also give an example of how to use selector interface to implement new customized selectors. The selection occurrs whenever the global model sends the weights to the local models, and the selected local trainers are stored as a list in selector which is also in channel.

New tasklets `increment_round` and `inform_end_of_training` are created to avoid the situation that the aggregator ends before trainers.

New class `MessageType` was created to pass different types of messages through channels.