feature: support ring-allreduce
ring-allreduce in distributed learning is implemented. The
implementation is based on baidu's allreduce implementation on
github.

This implementation doesn't rely on mpi or other backend. It's built
on top of channel - a concept in the flame project.

Unlike the standard protocol, this one allows a worker to join
distributed learning on a per-round fashion. In case there is no
agreement on membership among workers, workers do local training for
the round and try to agree on them in the next round. When there is an
agreement, then they start to conduct ring-allreduce.

Currently, worker departure (drop) is not supported. If such a case
happens, all workers will fail.