feature: hybrid mode with ring allreduce
Hybrid mode of combining distributed learning and federated learning
is implemented. This implementation is based on PR #131.

The current implementation has a bug thta can cause a deadlock when
trainers arrive late. Addressing the issue will be handled as a
separate PR as it requires quite changes in backend, channel, channel
manager, etc.