fix: message mixed up in distributed learning mode (#167)
During the member check, there is a high chance to receive weights
message because other workers started ring all-reduce process while
a trainer is in a member check loop. In such a case, the current code
will delcare failure of the membership check. This creates an
inconsistent agreement among workers, meaning that some think they are
in a ring and some others don't. To address the issue, we reinstate
the peek statement.

In addition, some cosmetric reformatting is done.