Mongodb impl for deployer registration
To complete the compute cluster registration, the controller uses mongodb to persist information. A cluster admin specifies computeId and apikey for the cluster and these are used during registration. If re-registration by same admin is attempted, no updates are made to the database. If new registration is attempted by different admin (different apikey) but same computeId, an error message is returned.