Duplicating controllers deployment code into deployer (#205)
This pr is a step towards implementing the core deployment logic into the deployer. We duplicate the relevant code and logic from the controller's deployer and put it as placeholder in the standalone deployer's code. At a later stage, only logic within the deployer will remain and all deployment related code from the controller will be removed.