chore: a skeleton script to configure minikube
This is a skeleton script to automate minikube configuration for fiab
on different operating systems and linux distributions.
Mac Os, Ubuntu and Amazon linux 2 will be supported in this automated
script. Note that users can manually configure minikube on different
linux distributions. The support for other linux distributions will be
planned in the future.