Deployer Registration Changes (#286)
Region is now called Realm. It is an array of tags used to describe geo-domain of the underlying cluster managed by the deployer. Additionally, property attribute is added which will be used in the future to provide cluster properties such as GPU/TPU, storage information etc.. Such properties will be used by the ML developer to provide fine-grain specification about the resources required to run their job.

Co-authored-by: Harshit <hd@Harshits-MacBook-Pro.local>