chore: automate minikube installation in amzn2 (#287)
The steps to install and configure minikube in amzn2 are tedious.
This script automates the installation and configuration in an amzn2
ec2 instance.

With amzn2 image and GPU instance, enabling custom domain (e.g.,
flame.test) appears very challenging. So, flame.sh outputs a message
to ask a user to set up the domain in route 52.

In addition, relevant documents are also renamed and revised.