fix: installation script for amzn2
In the script, crictl is installed in /usr/local/bin. However,
when minikube is executed with sudo, minikube fails to findcrictl.
The path needs to be enabled either sudoer's configuration or in
root's path, each of which may not be ideal. As a workaround, a
symbolic link is created under /usr/bin.

Also, during the testing, flame.sh has a bug of failing to run jq
correctly when ./flame.sh stop is executed. This is because the jq's
verions is < 1.6. The fix (adding '.') was applied.