doc: fiab guide revision
The fiab guide is outdated and incorrect. The system requirement for
the env is updated. A note on command "minikube delete " was
misplaced, thus causing confusion. The location of the note is
updated. Finally, an image pull from minikube may fail due to a
timeout (the default timeout value of 2 minutes, which is hard to
change). A workaround for the problem is documented.