feddyn implementation pytorch
Added FedDyn optimizer implementation (as described in the paper's pseudocode) compatible with pytorch only.

This involves an update function called by horizontal/trainer.py on the regularizers, and the trainer main.py files must send a deepcopy of the initial model (at each round) and the current model into the regularizer to get the algorithm's loss term.
The FedDyn regularizer also stores some information from the trainers, isolating the overhead incurred (from deepcopying a model) to just that optimizer.