feat+refactor+fix: per-channel backend support in SDK
The config format for per-channel backend is revised such that it is
well structured. Along with the format change, the logic to consume
per-channel backend in the ChannelManager is refactored due to the
config format changes.

In the testing with the hybrid example, a bug manifested in the mqtt
broker. If a message arrives after _rx_task is cancelled, setting a
message in the asyncio's future raises an exception. As a fix, if the
future is cancelled, the function on_message simply returns.