fix: hardening coordinated syncfl (#385)
The current implementation of the coordinated syncfl has some corner
cases.

1) Channel property for round is not specified; this causes the
default selector to ignore ends (peers in the other end of channel)
who join late.
2) There exists a synchronization issue; the coordinator returns
peer information so that middle aggregators and trainers can talk to
each other. However, the peer information may not be up to date yet
because the actual joining process to the channel between middle
aggregators and trainers are not completely.

These are handled in this change.