updated feddyn implementation pytorch
The new implementation matches the algorithm implemented in the paper.
This is no longer is similar to the pseudocode they provide.

The new implmentation includes an updated communication protocol where all trainers send the dataset size for that round to the aggregator at the beginning of every round.
This allows the aggregator to compute adaptive hyperparameters for each trainer individually and then send ALPHA_ADPT along with the global model to begin training for the round.