feat: optimize composer's extensibility (#390)
When a class is inherited, the current compose() function doesn't
reuse the tasklets created from the parent class. Therefore, a
developer needs to create many of the same tasklets so that they can
be chained (hence, composed). This duplicates the code.

To avoid this redundancy, the composer is extended such that the
parent's tasklets are retrieved with alias used when they were created
in the parent class.

To demonstrate this feature, the compose() method of top_aggregator in
asyncfl mode is modified. All other examples still follow the old way
of duplicating the code. Those can be revised over time.