feat: coordination revision
The coordinated FL is resived. The key changes are:

1) The coordinator has full responsibility of informing the end
of the training to the workers of all roles (top agg, mid agg and
trainer).

2) The top aggregator now receives from the coordinator a list of
middle aggregators that it can interact with. This can prevent the top
aggregator from sending a global model to the middle aggregators which
don't have any trainers. This improves efficiency by avoiding
unnecessary data communication. It may help to support serverless fl
case.

3) For (1), the coordinator expects the rounds information in its
config file.

4) The coordinator and middle aggregator can exchange meta information.
The middle aggregator can share extra information for its load or
resource utilization, etc. This kind of information can be utilized to
make the coordinator make a smart decision on mapping middle
aggregators with trainers.