feat: channel leave functionality for p2p backend (#398)
A functionality to leave a channel is not supported so far. The
functionality can be useful when a worker needs to leave and rejoin a
channel based on some conditions. To support leave functionality,
resource deallocaiton and cleanup of internal state in the
backend are implemented.

Note that channel leave functionality for mqtt backend is not
complete, which will be done separately in a future PR.