feddyn algorithm correction
The FedDyn algorithm does not send out the same model for training as it uses for evaluation.
This is now implemented in this PR.

The labels will be changed to type long before being sent to the loss function in order to avoid NaN values as output.

Lastly, the FedDynRegularizer class now has a .to(device) function in order to transfer the information as needed for subprocesses that use the GPU.