fix: slow tx task termination
tx task termination is slow because putting EMPTY_PAYLOAD was never
executed. This was because channel.remove(end) is called that end is
going to be removed. Hence, the length of channel._ends is always
zero. To address the issue, in channel.remove(), puting EMPTY_PAYLOAD
into the tx queue is implemented. With this, every time
channel.remove() is called, the given end's tx task is safely
terminated.