track tasklet runtime
Using a decorator created in flame/monitor/util.py, tasklets now can send runtime information to a MetricCollector in a composer.
The composer is initiliazed only once per process in the Role abstract class.
This allows for all runtime metrics to be stored in the same MetricCollector instance within a process.

Additionally, some imports in flame/mode have been adjusted, and aliases for feddyn were added as well.