feat: utilization/memory usage on CPU and GPU
The metric collector now runs two additional threads to track GPU/CPU utilization and memory usage.
So far, GPU utilization is still not specific to the process that is running.

Additionally, the metric collector was refactored in order to store samples in a variable stat_log.
Names of metrics now use a period to separate the metric type and alias.

setup.py has been updated to include gpustat and psutil.